flask_part/requirements.txt
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.5.0
requests==2.28.1
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
